% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/el_topo.R
\name{el_topo}
\alias{el_topo}
\title{El Topo colour palette function. Each palette has 5 or 6 colours given in rgb format.}
\usage{
el_topo(num)
}
\arguments{
\item{input}{Which of the 4 palettes from El Topo are wanted. 1:4.}
}
\description{
El Topo colour palette function. Each palette has 5 or 6 colours given in rgb format.
}
\examples{
el_topo(1)
}

