% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/holy_mountain.R
\name{holy_mountain}
\alias{holy_mountain}
\title{Holy_mountain colour palette function.}
\usage{
holy_mountain(num)
}
\arguments{
\item{input}{Which of the 7 palettes from The Holy Mountain are wanted. 1:7.}
}
\description{
Holy_mountain colour palette function.
}
\examples{
holy_mountain(1)
}

