% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/santa_sangre.R
\name{santa_sangre}
\alias{santa_sangre}
\title{Santa Sangre colour palette function.}
\usage{
santa_sangre(num)
}
\arguments{
\item{input}{Which of the 8 palettes from Santa Sangre are wanted. 1:8.}
}
\description{
Santa Sangre colour palette function.
}
\examples{
santa_sangre(1)
}

